Improve connection empty state UI
